pyproject.toml
[project]
name = "pytest-template"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pytest>=8.4.1",
    "pytest-check>=2.5.3",
    "pytest-cov>=6.2.1",
    "pytest-xdist>=3.7.0",
    "requests>=2.32.4",
    "loguru>=0.7.3",
    "pre-commit>=4.2.0",
    "allure-pytest>=2.14.3",
    "pydantic>=2.11.7",
]

[tool.pytest.ini_options]
addopts = "-ra -q"
testpaths = [
    "tests",
]

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true
